requirements.txt
pyttsx3>=2.90
SpeechRecognition>=3.8.1
pyaudio>=0.2.11
pyautogui>=0.9.53
psutil>=5.9.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
gTTS>=2.3.2
playsound>=1.3.0
